plum_backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
slack-sdk>=3.27.2
twilio>=9.0.5
anthropic>=0.28.0
python-multipart>=0.0.9
aiofiles>=23.2.1
httpx>=0.27.0
psycopg2-binary>=2.9.9
